model_backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.12
gitpython==3.1.44
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typer==0.15.2
typing-extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.96
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
